Fixed a bug with printing german umlaute

+++ lib/c/stdio/bputc.c
@@ -38,5 +38,7 @@ s32 bputc(FILE *f,char c) {
 			return EOF;
 		buf->buffer[buf->pos++] = c;
 	}
-	return c;
+	/* the cast is necessary to ensure that chars that are < 0 (e.g. german umlaute), won't be
+	 * converted to a s32 < 0, but to a s32 > 0 with the 8 LSB's filled */
+	return (u8)c;
 }
